Added exhibits -- both html and png
Interacted plots -- both scatterplots and parallel coordinate -- to examine the correlation between conjectured variables and COVID death rates by country. The COVID death rate data (total deaths by total cases) are as of 05-01-2021. The variables examined are the following: 
1) Positive Test Rate 
2) Obesity Preevalence
3) Median Age
4) Healthcare Expenditure per Individual
5) Hospital Capacity (hospital beds per individual)

The parallel plots are meant to show the entirety of all the examined variables by country.